requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-dotenv==1.0.0
transformers
torch
sentencepiece
openai
